+++ src/borg/conftest.py
@@ -162,7 +162,7 @@ class ArchiverSetup:
         self.patterns_file_path: str | None = None
 
     def get_kind(self) -> str:
-        if self.repository_location.startswith("rest://"):
+        if self.repository_location.startswith("ssh://"):
             return "remote"
         elif self.EXE == "borg.exe":
             return "binary"
@@ -218,7 +218,7 @@ def archiver(tmp_path, set_env_variables):
 
 @pytest.fixture()
 def remote_archiver(archiver, monkeypatch):
-    archiver.repository_location = "rest://" + "/" + str(archiver.repository_path)
+    archiver.repository_location = "ssh://__testsuite__/" + str(archiver.repository_path)
     # don't measure coverage in the "borg serve --rest" children, see #9470
     monkeypatch.delenv("COVERAGE_PROCESS_CONFIG", raising=False)
     monkeypatch.delenv("COVERAGE_PROCESS_START", raising=False)